Created the orientation function
Returns an array of the commented float values

+++ Main_code_Mk_1.0/Burnout.ino
@@ -3,7 +3,7 @@ int burnout(){
   int nextState = 4;
 
   float alts = altSensor.getAlt();
-  int orient = orientation();
+  float orient = orientation();
 
   currentAlt = alts[0];
   lastAlt = alts[1];
